SWDEV-312288 - Fix package licensing
The Debian packaging spec requires a copyright file to be installed into

/usr/share/doc/PACKAGE_NAME/copyright

This should be a machine readable file with all applicable licenses and
copyrights. Unless there's any complicated licensing, using only the
license file should be sufficient.

RPM only requires the LICENSE field to be set.

Note: LICENSE.TXT was incorrectly named, so I fixed it as this logic
depends on knowing the filename.

Signed-off-by: Jeremy Newton <Jeremy.Newton@amd.com>
Change-Id: I35b4af7683836e636e099da608fca2d46d63b40d