SWDEV-439965: Changing occupancy format from a value per CU to an event per wave slot.
Change-Id: I2b5b88e850c2618486d652f6e94a4340c006ab02

This method will allow the offline viewer to display wave start/end in a similar format to the Thread Trace Viewer without requiring the waves to be on the target_cu.
The occupancy event will now carry information about which SIMD and WaveSlot the wave occupies. Occupancy on public version has been temporarily removed.